调整调度方式

+++ Demo/Keil-ARM/stm32f4xx/Src/Func/SysTick.h
@@ -20,7 +20,7 @@
 extern s8 SysTick_us_Init_Test_Flag;
 
 
-s8 SysTick_us_Init(u8 SysCLK,u16 ms);
+s8 SysTick_us_Init(u8 SysCLK,u32 ms);
 void SysTick_us(u32 us);
 #endif
